# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
security:
  - ClerkBearerAuth: []
...
paths:
...
  /classes:
    get:
      operationId: listClasses
      tags: [classes]
      summary: List all gym classes
      description: Returns all upcoming gym classes for the authenticated caller's approved gym access
      responses:
        "200":
          description: List of gym classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GymClass"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/enrolled:
    get:
      operationId: listEnrolledClassIds
      tags: [classes]
      summary: List the caller's enrolled class ids
      description: Returns upcoming class ids that the authenticated member is enrolled in
      responses:
        "200":
          description: Enrolled class ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnrolledClassIdsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/waitlisted:
    get:
      operationId: listWaitlistedClassIds
      tags: [classes]
      summary: List the caller's waitlisted class ids
      description: Returns upcoming class ids that the authenticated member has waitlisted
      responses:
        "200":
          description: Waitlisted class ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistedClassIdsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/{id}/enroll:
    post:
      operationId: enrollInClass
      tags: [classes]
      summary: Enroll the caller in a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class enrollment state after enrolling
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
        "400":
          description: Invalid class id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Class is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: leaveClass
      tags: [classes]
      summary: Remove the caller from a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class enrollment state after leaving
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
        "400":
          description: Invalid class id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classes/{id}/waitlist:
    post:
      operationId: joinClassWaitlist
      tags: [classes]
      summary: Join the waitlist for a full class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class waitlist state after joining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
        "400":
          description: Invalid class id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Class cannot be waitlisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: leaveClassWaitlist
      tags: [classes]
      summary: Leave the waitlist for a class
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Class waitlist state after leaving
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GymClass"
        "400":
          description: Invalid class id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Class not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ClerkBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Clerk session bearer token sent as Authorization: Bearer <token>."
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: ["string", "null"]
      required:
        - error
...
    EnrolledClassIdsResponse:
      type: object
      properties:
        classIds:
          type: array
          items:
            type: string
      required:
        - classIds
    WaitlistedClassIdsResponse:
      type: object
      properties:
        classIds:
          type: array
          items:
            type: string
      required:
        - classIds
    GymClass:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum: [Yoga, Zumba, CrossFit, HIIT, Spinning, Boxing, Pilates, Strength, Cardio, Other]
        description:
          type: string
        trainer:
          type: string
        date:
          type: string
        startTime:
          type: string
        duration:
          type: integer
        maxParticipants:
          type: integer
        enrolledCount:
          type: integer
        waitlistedCount:
          type: integer
        checkedInCount:
          type: integer
        room:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        color:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - category
        - description
        - trainer
        - date
        - startTime
        - duration
        - maxParticipants
        - enrolledCount
        - waitlistedCount
        - checkedInCount
        - room
        - status
        - color
        - createdAt
        - updatedAt
...